backend/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==4.2.20
django-cors-headers==4.7.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pycparser==2.22
PyJWT==2.9.0
six==1.17.0
sqlparse==0.5.3
pytest==8.3.5
freezegun==1.5.1
pytest-django==4.11.1
pytest-cov==6.1.1
dj_rest_auth==7.0.1
